requirements.txt
mysqlclient==1.3.12
jsonfield==1.0.3
six==1.10.0
Django==2.0.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
redis==2.10.6
gevent==1.2.2
gunicorn==19.7.1
